+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - GP2-1068 - Autoformat Python code with black and isort
 - GP2-1111 - EP responsive marketing approach changes
 - GP2-1029 - Population data sources attributions and links
+- GP2-974 - target market research page
 
 ### Fixed bugs